pyproject.toml
[tool.poetry]
name = "blog-api"
version = "0.1.0"
description = ""
authors = ["oianshek <oiansheks@gmai.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.115.4"
pydantic-settings = "^2.6.0"
uvicorn = "^0.32.0"
asyncpg = "^0.30.0"
sqlalchemy = "^2.0.36"
passlib = "^1.7.4"
python-jose = "^3.3.0"
alembic = "^1.13.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
